Added extra options to each reminder. Make preparation for integrating shopping lists to app.

+++ app/src/main/java/com/cray/software/justreminder/datas/EventsDataProvider.java
@@ -137,7 +137,7 @@ public class EventsDataProvider {
                 int myMonth = s.getInt(s.getColumnIndex(Constants.COLUMN_MONTH));
                 int myYear = s.getInt(s.getColumnIndex(Constants.COLUMN_YEAR));
                 int repCode = s.getInt(s.getColumnIndex(Constants.COLUMN_REPEAT));
-                int remCount = s.getInt(s.getColumnIndex(Constants.COLUMN_REMINDERS_COUNT));
+                long remCount = s.getLong(s.getColumnIndex(Constants.COLUMN_REMINDERS_COUNT));
                 int isDone = s.getInt(s.getColumnIndex(Constants.COLUMN_IS_DONE));
                 long afterTime = s.getInt(s.getColumnIndex(Constants.COLUMN_REMIND_TIME));
                 String mType = s.getString(s.getColumnIndex(Constants.COLUMN_TYPE));
